Consolidating Inbox and Outbox Pages into Messages; Update routing to /PASS/messages; Created internal tab to toggle between Inbox and Outbox components in Messages; Refactored Inbox and Outbox components into MessageFolders; Include corrections to JSDoc; Moved all styled-components styling into MessageStyles; Minor edit to MessageContext to setLoadingMessage to true before fetching data
