+++ src/AppRoutes.jsx
@@ -1,18 +1,22 @@
 // React Imports
 import React, { useState, useEffect } from 'react';
 import { Routes, Route, Navigate, Outlet } from 'react-router-dom';
-// Page Imports
+// Inrupt Imports
 import { useSession } from '@inrupt/solid-ui-react';
+// Custom Hook Imports
 import { useRedirectUrl } from './hooks';
-
-import Home from './routes/Home';
-import Clients from './routes/Clients';
-import Inbox from './routes/Inbox';
-import Outbox from './routes/Outbox';
-import Forms from './routes/Forms';
+// Page Imports
+import { Home, Clients, Messages, Forms } from './routes';
 
 const ProtectedRoute = ({ isLoggedIn, children }) =>
   isLoggedIn ? children ?? <Outlet /> : <Navigate to="/PASS/" replace />;
+
+/**
+ * The main application routing for PASS
+ *
+ * @name AppRoutes
+ * @returns {React.JSX.Element}
+ */
 
 const AppRoutes = () => {
   const { session } = useSession();
@@ -49,8 +53,7 @@ const AppRoutes = () => {
       <Route element={<ProtectedRoute isLoggedIn={session.info.isLoggedIn} />}>
         <Route path="/PASS/clients" element={<Clients />} />
         <Route path="/PASS/forms" element={<Forms />} />
-        <Route path="/PASS/inbox" element={<Inbox />} />
-        <Route path="/PASS/outbox" element={<Outbox />} />
+        <Route path="/PASS/messages" element={<Messages />} />
         <Route path="*" element={<Navigate to={restorePath} replace />} />
       </Route>
     </Routes>
