+++ src/components/Messages/MessagePreview.jsx
@@ -1,7 +1,17 @@
 // React imports
 import React, { useState } from 'react';
 // Styling Imports
-import styled from 'styled-components';
+import { StyledDate, StyledHeader, StyledPreview } from './MessageStyles';
+
+/**
+ * messageFolderProps is an object that stores the props for the MessageFolder
+ * component
+ *
+ * @typedef messageFolderProps
+ * @type {object}
+ * @property {string} message - The content of the message sent
+ * @memberof typedefs
+ */
 
 /**
  * Message Preview Component - Component that displays message previews from
@@ -9,6 +19,7 @@ import styled from 'styled-components';
  *
  * @memberof Inbox
  * @name MessagePreview
+ * @param {messageFolderProps} Props - Component props for MessagePreview
  */
 
 const MessagePreview = ({ message }) => {
@@ -28,28 +39,5 @@ const MessagePreview = ({ message }) => {
     </StyledPreview>
   );
 };
-
-const StyledPreview = styled('div')({
-  display: 'flex',
-  flexDirection: 'column',
-  border: '2px solid black',
-  borderRadius: '10px',
-  padding: '5px',
-  cursor: 'pointer',
-  '&:hover': {
-    boxShadow: '2px 3px 2px rgba(0, 0, 0, .4)'
-  }
-});
-
-const StyledDate = styled('p')({
-  margin: 0,
-  padding: 0
-});
-
-const StyledHeader = styled('h1')({
-  margin: 0,
-  padding: 0,
-  fontSize: '1.2rem'
-});
 
 export default MessagePreview;
